Refactor ubication components
