+++ src/features/ubication/constants.ts
@@ -1,0 +1,8 @@
+export const FREEZER_RESET_WARNING =
+  "¿Estás seguro que la ubicación no es un freezer?\n" +
+  "Si tienes alimentos en periodo de congelado su progreso se reiniciará";
+
+export const DELETE_UBICATION_WARNING =
+  "¿Estás seguro que quieres eliminar esta ubicación?\n" +
+  "Todos los alimentos que se encuentren en el se borrarán y no podrán ser recuperados.\n" +
+  "Cámbialos de ubicación primero si deseas conservarlos";
